src/main/java/com/ninjaone/rmmrestapi/controller/ServiceController.java
package com.ninjaone.rmmrestapi.controller;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import com.ninjaone.rmmrestapi.model.Service;
import com.ninjaone.rmmrestapi.service.ServiceService;

@RestController
@RequestMapping("/service")
public class ServiceController {
    private final ServiceService service;

    public ServiceController(ServiceService service) {
        this.service = service;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    private Service post(@RequestBody Service entity) {
        return service.save(entity);
    }
    
    @GetMapping
    private List<Service> get() {
        return service.getAll();
    }

    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    private void deleteServiceEntity(@PathVariable Integer id) {
        service.deleteById(id);
    }
}
